finishes deleteBet

+++ controllers/bets.js
@@ -46,19 +46,15 @@ const updateBet = async (req, res) => {
 }
 
 const deleteBet = async (req, res) => {
-  try {
-    const { id: betID } = req.params
-    if (!mongoose.Types.ObjectId.isValid(betID)) {
-      return res.status(400).json({ msg: `Invalid ID format: ${betID}` })
-    }
-    const bet = await Bet.findOneAndDelete({ _id: betID })
-    if (!bet) {
-      return res.status(404).json({ msg: `No bet with ID: ${betID}` })
-    }
-    res.status(200).json({ msg: 'Bet deleted' })
-  } catch (error) {
-    res.status(500).json({ msg: error })
+  const {
+    user: { userID },
+    params: { id: betID },
+  } = req
+  const bet = await Bet.findOneAndDelete({ _id: betID, createdBy: userID })
+  if (!bet) {
+    throw new NotFoundError(`No bet with ID: ${betID}`)
   }
+  res.status(StatusCodes.OK).send()
 }
 
 module.exports = { getAllBets, createBet, getBet, updateBet, deleteBet }
